fix(dropdown): API query does not run when the down arrow key is used to open
When the dropdown utilizes API settings, it retrieves data from a remote
source only if there are no items present when the dropdown is opened.
The dropdown module removes any existing items if it is activated by
clicking the icon or focusing on the search field.

However, when the down arrow key is pressed to open the dropdown, the
existing items, which are designated as initial values, remain intact,
preventing API query from being executed.

This issue is addressed by assigning the global variable `focused`
a value of `true` when the down arrow key is pressed.